add indicators for speed, alt und vertical speed. add Ability to hide indicators

+++ src/example/MainWindow.cpp
@@ -83,7 +83,7 @@ MainWindow::MainWindow( QWidget *parent ) :
 
 MainWindow::~MainWindow()
 {
-    cout << "Average time step: " << ( (double)m_realTime ) / ( (double)m_steps ) << " s" << endl;
+    cout << "Average time step: " << ( static_cast<double>(m_realTime) ) / ( static_cast<double>(m_steps) ) << " s" << endl;
 
     if ( m_timerId ) killTimer( m_timerId );
 
@@ -98,45 +98,45 @@ void MainWindow::timerEvent( QTimerEvent *event )
     QMainWindow::timerEvent( event );
     /////////////////////////////////
 
-    float timeStep = m_time.restart();
+    double timeStep = m_time.restart();
 
-    m_realTime = m_realTime + timeStep / 1000.0f;
+    m_realTime = m_realTime + timeStep / 1000.0;
 
-    float alpha     =  0.0f;
-    float beta      =  0.0f;
-    float roll      =  0.0f;
-    float pitch     =  0.0f;
-    float heading   =  0.0f;
-    float slipSkid  =  0.0f;
-    float turnRate  =  0.0f;
-    float devH      =  0.0f;
-    float devV      =  0.0f;
-    float airspeed  =  0.0f;
-    float altitude  =  0.0f;
-    float pressure  = 28.0f;
-    float climbRate =  0.0f;
-    float machNo    =  0.0f;
-    float adf       =  0.0f;
-    float dme       =  0.0f;
+    double alpha     =  0.0;
+    double beta      =  0.0;
+    double roll      =  0.0;
+    double pitch     =  0.0;
+    double heading   =  0.0;
+    double slipSkid  =  0.0;
+    double turnRate  =  0.0;
+    double devH      =  0.0;
+    double devV      =  0.0;
+    double airspeed  =  0.0;
+    double altitude  =  0.0;
+    double pressure  = 28.0;
+    double climbRate =  0.0;
+    double machNo    =  0.0;
+    double adf       =  0.0;
+    double dme       =  0.0;
 
     if ( m_ui->pushButtonAuto->isChecked() )
     {
-        alpha     =   20.0f * sin( m_realTime /  10.0f );
-        beta      =   15.0f * sin( m_realTime /  10.0f );
-        roll      =  180.0f * sin( m_realTime /  10.0f );
-        pitch     =   90.0f * sin( m_realTime /  20.0f );
-        heading   =  360.0f * sin( m_realTime /  40.0f );
-        slipSkid  =    1.0f * sin( m_realTime /  10.0f );
-        turnRate  =    7.0f * sin( m_realTime /  10.0f );
-        devH      =    1.0f * sin( m_realTime /  20.0f );
-        devV      =    1.0f * sin( m_realTime /  20.0f );
-        airspeed  =  125.0f * sin( m_realTime /  40.0f ) +  125.0f;
-        altitude  = 9000.0f * sin( m_realTime /  40.0f ) + 9000.0f;
-        pressure  =    2.0f * sin( m_realTime /  20.0f ) +   30.0f;
-        climbRate =  650.0f * sin( m_realTime /  20.0f );
-        machNo    = airspeed / 650.0f;
-        adf       = -360.0f * sin( m_realTime /  50.0f );
-        dme       =   99.0f * sin( m_realTime / 100.0f );
+        alpha     =   20.0 * sin( m_realTime /  10.0 );
+        beta      =   15.0 * sin( m_realTime /  10.0 );
+        roll      =  180.0 * sin( m_realTime /  10.0 );
+        pitch     =   90.0 * sin( m_realTime /  20.0 );
+        heading   =  360.0 * sin( m_realTime /  40.0 );
+        slipSkid  =    1.0 * sin( m_realTime /  10.0 );
+        turnRate  =    7.0 * sin( m_realTime /  10.0 );
+        devH      =    1.0 * sin( m_realTime /  20.0 );
+        devV      =    1.0 * sin( m_realTime /  20.0 );
+        airspeed  =  125.0 * sin( m_realTime /  40.0 ) +  125.0;
+        altitude  = 9000.0 * sin( m_realTime /  40.0 ) + 9000.0;
+        pressure  =    2.0 * sin( m_realTime /  20.0 ) +   30.0;
+        climbRate =  650.0 * sin( m_realTime /  20.0 );
+        machNo    = airspeed / 650.0;
+        adf       = -360.0 * sin( m_realTime /  50.0 );
+        dme       =   99.0 * sin( m_realTime / 100.0 );
 
         m_ui->spinBoxAlpha ->setValue( alpha     );
         m_ui->spinBoxBeta  ->setValue( beta      );
@@ -157,29 +157,29 @@ void MainWindow::timerEvent( QTimerEvent *event )
     }
     else
     {
-        alpha     = (float)m_ui->spinBoxAlpha ->value();
-        beta      = (float)m_ui->spinBoxBeta  ->value();
-        roll      = (float)m_ui->spinBoxRoll  ->value();
-        pitch     = (float)m_ui->spinBoxPitch ->value();
-        heading   = (float)m_ui->spinBoxHead  ->value();
-        slipSkid  = (float)m_ui->spinBoxSlip  ->value();
-        turnRate  = (float)m_ui->spinBoxTurn  ->value();
-        devH      = (float)m_ui->spinBoxDevH  ->value();
-        devV      = (float)m_ui->spinBoxDevV  ->value();
-        airspeed  = (float)m_ui->spinBoxSpeed ->value();
-        pressure  = (float)m_ui->spinBoxPress ->value();
-        altitude  = (float)m_ui->spinBoxAlt   ->value();
-        climbRate = (float)m_ui->spinBoxClimb ->value();
-        machNo    = (float)m_ui->spinBoxMach  ->value();
-        adf       = (float)m_ui->spinBoxADF   ->value();
-        dme       = (float)m_ui->spinBoxDME   ->value();
+        alpha     = static_cast<double>(m_ui->spinBoxAlpha) ->value();
+        beta      = static_cast<double>(m_ui->spinBoxBeta)  ->value();
+        roll      = static_cast<double>(m_ui->spinBoxRoll)  ->value();
+        pitch     = static_cast<double>(m_ui->spinBoxPitch) ->value();
+        heading   = static_cast<double>(m_ui->spinBoxHead)  ->value();
+        slipSkid  = static_cast<double>(m_ui->spinBoxSlip)  ->value();
+        turnRate  = static_cast<double>(m_ui->spinBoxTurn)  ->value();
+        devH      = static_cast<double>(m_ui->spinBoxDevH)  ->value();
+        devV      = static_cast<double>(m_ui->spinBoxDevV)  ->value();
+        airspeed  = static_cast<double>(m_ui->spinBoxSpeed) ->value();
+        pressure  = static_cast<double>(m_ui->spinBoxPress) ->value();
+        altitude  = static_cast<double>(m_ui->spinBoxAlt)   ->value();
+        climbRate = static_cast<double>(m_ui->spinBoxClimb) ->value();
+        machNo    = static_cast<double>(m_ui->spinBoxMach)  ->value();
+        adf       = static_cast<double>(m_ui->spinBoxADF)   ->value();
+        dme       = static_cast<double>(m_ui->spinBoxDME)   ->value();
     }
 
     m_ui->widgetPFD->setFlightPathMarker ( alpha, beta );
     m_ui->widgetPFD->setRoll          ( roll     );
     m_ui->widgetPFD->setPitch         ( pitch     );
     m_ui->widgetPFD->setSlipSkid      ( slipSkid  );
-    m_ui->widgetPFD->setTurnRate      ( turnRate / 6.0f );
+    m_ui->widgetPFD->setTurnRate      ( turnRate / 6.0 );
     m_ui->widgetPFD->setDevH          ( devH      );
     m_ui->widgetPFD->setDevV          ( devV      );
     m_ui->widgetPFD->setHeading       ( heading   );
@@ -187,13 +187,13 @@ void MainWindow::timerEvent( QTimerEvent *event )
     m_ui->widgetPFD->setMachNo        ( machNo    );
     m_ui->widgetPFD->setAltitude      ( altitude  );
     m_ui->widgetPFD->setPressure      ( pressure  );
-    m_ui->widgetPFD->setClimbRate     ( climbRate / 100.0f );
+    m_ui->widgetPFD->setClimbRate     ( climbRate / 100.0 );
     m_ui->widgetPFD->setIdentifier    ( "ILAX" , true );
     m_ui->widgetPFD->setDistance      ( dme    , true );
 
     m_ui->widgetNAV->setHeading    ( heading   );
-    m_ui->widgetNAV->setHeadingBug ( 0.0f );
-    m_ui->widgetNAV->setCourse     ( 0.0f );
+    m_ui->widgetNAV->setHeadingBug ( 0.0 );
+    m_ui->widgetNAV->setCourse     ( 0.0 );
     m_ui->widgetNAV->setBearing    ( adf  , true );
     m_ui->widgetNAV->setDeviation  ( devH , true );
     m_ui->widgetNAV->setDistance   ( dme  , true );
@@ -204,7 +204,7 @@ void MainWindow::timerEvent( QTimerEvent *event )
     m_ui->widgetSix->setPressure  ( pressure  );
     m_ui->widgetSix->setAirspeed  ( airspeed  );
     m_ui->widgetSix->setHeading   ( heading   );
-    m_ui->widgetSix->setSlipSkid  ( slipSkid * 15.0f );
+    m_ui->widgetSix->setSlipSkid  ( slipSkid * 15.0 );
     m_ui->widgetSix->setTurnRate  ( turnRate  );
     m_ui->widgetSix->setClimbRate ( climbRate );
